only showing movies that match params[:ratings]
Uses a new function in the model (movies_with_ratings).

There must be a way to sort the movies in the model (while only grabbing
ones that match a certain condition. I didn't particularly want to pass
in my sort_by value to the model (it would have made for a complicated
function passing) - it felt cleaner to do it in the controller. The
methods for selecting from the database and sorting in the controller
can probably be cleaned up too.